backend/pyproject.toml
[project]
name = "ethglobal-cannes2025"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "fastapi[standard] (>=0.115.14,<0.116.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "web3 (>=7.12.0,<8.0.0)",
    "aiohttp (>=3.12.13,<4.0.0)",
    "aleph-sdk-python (>=2.0.5,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
repository = "https://github.com/RezaRahemtola/ETHGlobal-Cannes2025"


[tool.poetry.group.dev.dependencies]
mypy = "^1.16.1"
ruff = "^0.12.2"
